Change supporters to use mustache templating
This actually caused a net increase of LOC. However, the logic is now less
confusing for what each level has access to. I should be able to simplify the
supporters.json if I wanted to, now.

Next to change templating:
preview-template.html
and make a standard template that has the html head, body nav, and body footer
then put supporters, preview, and index into that.