Generate the RSS feeds from a template
This is a bit faster than using the feed library.
The issue was the feed library didn't support the isPermaLink attribute for the guid.
So I would have to go back to using a link in the guid, which feels wrong :P

I think this will help reduce code when I improve the scraper API.
(If you don't count the template as code.)

I'm not sure if this is the right decision, but I think it is OK.